+++ ConcernsCaseWork/Service.Redis/Models/CaseWrapper.cs
@@ -1,7 +1,6 @@
 ﻿using Service.TRAMS.Cases;
 using System;
 using System.Collections.Generic;
-using System.Numerics;
 
 namespace Service.Redis.Models
 {
@@ -9,6 +8,6 @@ namespace Service.Redis.Models
 	public sealed class CaseWrapper
 	{
 		public CaseDto CaseDto { get; set; }
-		public IDictionary<BigInteger, RecordWrapper> Records { get; set; } = new Dictionary<BigInteger, RecordWrapper>();
+		public IDictionary<long, RecordWrapper> Records { get; set; } = new Dictionary<long, RecordWrapper>();
 	}
 }
