+++ ConcernsCaseWork/Service.Redis/Models/UserState.cs
@@ -1,6 +1,5 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.Numerics;
 
 namespace Service.Redis.Models
 {
@@ -8,6 +7,6 @@ namespace Service.Redis.Models
 	public sealed class UserState
 	{
 		public string TrustUkPrn { get; set; }
-		public IDictionary<BigInteger, CaseWrapper> CasesDetails { get; } = new Dictionary<BigInteger, CaseWrapper>();
+		public IDictionary<long, CaseWrapper> CasesDetails { get; } = new Dictionary<long, CaseWrapper>();
 	}
 }
