+++ ConcernsCaseWork/Service.Redis/RecordRatingHistory/IRecordRatingHistoryCachedService.cs
@@ -1,11 +1,10 @@
 ﻿using Service.TRAMS.RecordRatingHistory;
-using System.Numerics;
 using System.Threading.Tasks;
 
 namespace Service.Redis.RecordRatingHistory
 {
 	public interface IRecordRatingHistoryCachedService
 	{
-		Task<RecordRatingHistoryDto> PostRecordRatingHistory(RecordRatingHistoryDto recordRatingHistoryDto, string caseworker, BigInteger caseUrn);
+		Task<RecordRatingHistoryDto> PostRecordRatingHistory(RecordRatingHistoryDto recordRatingHistoryDto, string caseworker, long caseUrn);
 	}
 }
